Fix Superior 6000S: timer uses hours, immediate timer remaining update, MCU version, humidity sensor, remove error sensor, add ESP version
- Timer number changed from minutes to hours (HA auto-represents as Xh Ym)
- Timer remaining immediately updates when timer is set (was delayed ~60s)
- Timer remaining is now a sensor (hours) instead of text sensor
- MCU version decoding added for Superior (tag 0x01)
- ESP version text sensor added to Superior YAML
- Error text sensor removed from Superior YAML
- Humidity sensor: 0 decimal places, percent unit
- All device YAMLs updated to use number sensors for timer
- publish_number/publish_sensor/on_number_command signatures changed to float

+++ components/levoit/levoit.h
@@ -38,13 +38,13 @@ class Levoit : public Component, public uart::UARTDevice {
   
   // called by switch, number,... when HA toggles it
   void on_switch_command(SwitchType type, bool state);
-  void on_number_command(NumberType type, uint32_t value);  
+  void on_number_command(NumberType type, float value);  
   void on_select_command(SelectType type, uint32_t value);
 
   // called by decoder when device reports status
   void publish_switch(SwitchType type, bool state);
-  void publish_number(NumberType type, uint32_t value);
-  void publish_sensor(SensorType type, uint32_t value);
+  void publish_number(NumberType type, float value);
+  void publish_sensor(SensorType type, float value);
   void publish_select(SelectType type, uint32_t value);
   void publish_text_sensor(TextSensorType type, const std::string &value);
   // binary sensor + button
